+++ CHANGES.rst
@@ -13,6 +13,7 @@ in progress
 - Upgrade to ``paho-mqtt==1.3.1``
 - Improve logging
 - Increase publishing interval for "func:sine" to 0.5 seconds
+- Improve MQTT broker connection handling and logging
 
 
 2017-07-15 0.11.0